+++ src/tree/tree.hpp
@@ -9,7 +9,7 @@
 class Tree
 {
   public:
-	Tree(std::vector<std::vector<int>> data)
+	Tree(std::vector<std::vector<double>> data)
 		: data(data) {}
 	~Tree()
 	{
@@ -20,26 +20,26 @@ class Tree
 		root = _build_tree(data);
 	}
 	const std::string to_string();
-	EndNode *classify(std::vector<int> row)
+	EndNode *classify(std::vector<double> row)
 	{
 		return _classify(row, root);
 	}
 
   private:
-	std::vector<std::vector<int>> data;
+	std::vector<std::vector<double>> data;
 	Node *root;
 
-	Node *_build_tree(std::vector<std::vector<int>> rows);
-	std::tuple<int, Question *> find_best_split(std::vector<std::vector<int>> rows);
-	double gini(std::vector<std::vector<int>> rows);
-	std::tuple<std::vector<std::vector<int>>,
-			   std::vector<std::vector<int>>>
-	partition(std::vector<std::vector<int>> rows,
+	Node *_build_tree(std::vector<std::vector<double>> rows);
+	std::tuple<double, Question *> find_best_split(std::vector<std::vector<double>> rows);
+	double gini(std::vector<std::vector<double>> rows);
+	std::tuple<std::vector<std::vector<double>>,
+			   std::vector<std::vector<double>>>
+	partition(std::vector<std::vector<double>> rows,
 			  Question *question);
-	double info_gain(std::vector<std::vector<int>> left,
-					 std::vector<std::vector<int>> right,
+	double info_gain(std::vector<std::vector<double>> left,
+					 std::vector<std::vector<double>> right,
 					 double current_uncertainty);
-	EndNode *_classify(std::vector<int> row, Node *node);
+	EndNode *_classify(std::vector<double> row, Node *node);
 };
 
 #endif
